Adjust cell velocities in 2 steps
First, interpolate (average) the adjacent-cell normal velocities to
faces; second, update the adjacent cell velocities. The reason for
breaking this up into two steps is that the correction needs to be
applied to both adjacent cells for each marked thin-body face. There
isn't any guarantee about the order that these will be applied or that
they will be on the same grid, so in the previous implementation the
second updated cell would have been updated with an incorrect
interpolated face velocity.